Advanced subgradient optimization observation #6:
Suppose that u[i] = 0 and G[i] < 0 then u[i] will remain zero after
update.
But a G^2 factor will been included in the calculation for the step size.
There seem little effective reason for this - after all we know that u[i] is not going to alter.
Hence we have found it helpful to adjust the subgradients before
calculating the step size using:
G[i] = 0, if u[i] = 0 and G[i] < 0.

Signed-off-by: Luis H. P. Mendes <luishpmendes@gmail.com>