Fixing fixSolution function
The previous implementation of fixSolution assumed that it was always
possible to remove a conflicting edge of the infeasible solution, but it
is not true because it may be possible that all the conflicting edges
are bridges or that, after removing an edge, there is no other
non-conflicting edge to reconnect the solution.
Besides that, it was assumed that it was always possible to fix the
solution.

Signed-off-by: Luis H. P. Mendes <luishpmendes@gmail.com>