Modifying the stepSize calculation in relaxLag2 function.
It is helpful to replace primalBoundValue in the equation for the step
size by 1.05*primalBoundValue.
The reasoning here is that, ideally, we would like primalBoundValue and
dualBoundValue to coincide. As dualBoundValue approaches
primalBoundValue, using the original equation for the step size implies
that stepSize gets smaller and smaller, so that we may experience many
iterations as dualBoundValue creeps closer and closer to
primalBoundValue. Replacing primalBoundValue by 1.05*primalBoundValue
avoids this problem.

Signed-off-by: Luis H. P. Mendes <luishpmendes@gmail.com>